RavenDB-15976 Simplified code for journal write

+++ src/Voron/Impl/Journal/IJournalWriter.cs
@@ -6,7 +6,7 @@ namespace Voron.Impl.Journal
 {
     public unsafe interface IJournalWriter : IDisposable
     {
-        TimeSpan Write(long posBy4Kb, byte* p, int numberOf4Kb);
+        TimeSpan Write(long posBy4Kb, byte* p, long numberOf4Kb);
         int NumberOfAllocated4Kb { get; }
         bool Disposed { get; }
         bool DeleteOnClose { get; set; }
